Fixing gap between tiles, bug with campfire, other bugs

+++ Assets/Scripts/Interactables/CellDoor.cs
@@ -20,7 +20,8 @@ public class CellDoor : Interactable
         {
             PlaySound();
             StartCoroutine(PlayAnimation());
-            GameManager.GetInstance().DeleteKey(keyNeeded);
+            if(!openWithLever)
+                GameManager.GetInstance().DeleteKey(keyNeeded);
         }
     }
